feat(jsonnet): switch back to google/go-jsonnet
Due to recent changes to google/go-jsonnet we can drop our dirty patches and use
the official compiler again!
