+++ go.mod
@@ -6,7 +6,7 @@ require (
 	github.com/Masterminds/semver v1.4.2
 	github.com/alecthomas/chroma v0.6.6
 	github.com/fatih/color v1.7.0
-	github.com/google/go-jsonnet v0.13.0
+	github.com/google/go-jsonnet v0.14.1-0.20190924111122-dc5c94ac78f0
 	github.com/kr/pretty v0.1.0 // indirect
 	github.com/pkg/errors v0.8.1
 	github.com/posener/complete v1.2.1
@@ -22,3 +22,7 @@ require (
 	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
 	gopkg.in/yaml.v2 v2.2.2
 )
+
+// replace github.com/google/go-jsonnet 21c00f1b9ebc856aabf72e2bdb818e1f5b3c11f7 => github.com/sh0rez/go-jsonnet export-children
+
+replace github.com/google/go-jsonnet v0.14.1-0.20190924111122-dc5c94ac78f0 => github.com/sh0rez/go-jsonnet v0.13.1-0.20190926190918-c4e27aa523e7
